chore(structure): reorganize into cmd pattern
