+++ cmd/cli/main.go
@@ -3,8 +3,8 @@ package main
 // =^_^=
 
 import (
-	"go.nirbar.in/nala/internal/db"
-	"go.nirbar.in/nala/internal/ui"
+	"go.nirbar.in/nala/internal/cli/db"
+	"go.nirbar.in/nala/internal/cli/ui"
 )
 
 func main() {
