+++ internal/cli/tools/add.go
