+++ internal/cli/tools/multiply.go
