+++ internal/cli/ui/events.go
